requirements.txt
django
djangorestframework
psycopg2-binary
black
drf-spectacular